chore(product): keep test suites under the max-lines cap (PRO-265)
The chord-ownership integration test pushed ComposerRichTextEditor.test.tsx
past 600 lines; it lives with the other PRO-265 behavior coverage in
ComposerFormatGuardPlugin.test.tsx, which already carries the same harness.
