+++ apps/packages/product-client/src/components/workspace/chat/input/ComposerFormatGuardPlugin.test.tsx
@@ -4,12 +4,14 @@ import { act, cleanup, fireEvent, render, waitFor } from "@testing-library/react
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import {
   $createParagraphNode,
+  $createRangeSelection,
   $createTextNode,
   $getRoot,
   $getSelection,
   $isElementNode,
   $isParagraphNode,
   $isRangeSelection,
+  $setSelection,
   PASTE_COMMAND,
   type LexicalEditor,
 } from "lexical";
@@ -17,6 +19,8 @@ import {
   ComposerRichTextEditor,
   type ComposerRichTextEditorProps,
 } from "#product/components/workspace/chat/input/ComposerRichTextEditor";
+import { SHORTCUTS } from "#product/config/shortcuts/registry";
+import { shouldDispatchKeyboardShortcut } from "#product/lib/domain/shortcuts/dispatch-policy";
 
 let originalRangeRectDescriptor: PropertyDescriptor | undefined;
 
@@ -189,6 +193,49 @@ describe("ComposerFormatGuardPlugin", () => {
       $getRoot().getAllTextNodes().every((textNode) => textNode.getFormat() === 0)
     ))).toBe(true);
     expect(onChange.mock.calls.at(-1)?.[0]).toBe("under typed");
+  });
+});
+
+describe("composer bold-chord ownership", () => {
+  it("cedes the sidebar's B chord to bold only while text is highlighted", async () => {
+    const harness = renderEditor({ value: "" });
+    await harness.ready();
+    act(() => resetText(harness.editor, "chord"));
+    const boldChordEvent = {
+      key: "b",
+      code: "KeyB",
+      metaKey: true,
+      ctrlKey: false,
+      shiftKey: false,
+      altKey: false,
+      defaultPrevented: false,
+      target: harness.root,
+    } as unknown as KeyboardEvent;
+
+    // Collapsed caret: the editor does not claim the chord.
+    expect(harness.root.hasAttribute("data-chat-composer-highlight")).toBe(false);
+    expect(shouldDispatchKeyboardShortcut(SHORTCUTS.toggleLeftSidebar, boldChordEvent))
+      .toBe(true);
+
+    act(() => {
+      harness.editor.update(() => {
+        const textNode = $getRoot().getAllTextNodes()[0]!;
+        const selection = $createRangeSelection();
+        selection.setTextNodeRange(textNode, 0, textNode, textNode.getTextContentSize());
+        $setSelection(selection);
+      }, { discrete: true });
+    });
+
+    // Highlighted text: the editor claims the chord and the key bolds it.
+    await waitFor(() => {
+      expect(harness.root.hasAttribute("data-chat-composer-highlight")).toBe(true);
+    });
+    expect(shouldDispatchKeyboardShortcut(SHORTCUTS.toggleLeftSidebar, boldChordEvent))
+      .toBe(false);
+    fireEvent.keyDown(harness.root, { key: "b", ctrlKey: true });
+    await waitFor(() => {
+      expect(harness.root.querySelector(".font-semibold")?.textContent).toBe("chord");
+    });
   });
 });
